Refactoring the framework . created a framework class with a bootstrap method that instanciates all directives and plugs them into their target HTML elements while injecting all the necessary dependencies . commented the code
